refactor: remove play2.4 and scala 2.11 support
