--- src/main/play_json_2.4/com/codacy/client/bitbucket/BodyWriteableWrapper.scala
@@ -1,5 +1,0 @@
-package com.codacy.client.bitbucket
-
-object BodyWriteableWrapper {
-  type BodyWriteable[A] = play.api.http.Writeable[A]
-}
