--- src/main/play_json_2.4/com/codacy/client/bitbucket/DefaultBodyWritables.scala
@@ -1,3 +1,0 @@
-package com.codacy.client.bitbucket
-
-object DefaultBodyWritables extends play.api.http.DefaultWriteables
