--- src/main/play_json_2.4/com/codacy/client/bitbucket/JsResultHelper.scala
@@ -1,7 +1,0 @@
-package com.codacy.client.bitbucket
-
-import play.api.data.validation.ValidationError
-
-object JsResultHelper {
-  def error(error: String): ValidationError = ValidationError(error)
-}
