--- src/main/play_json_2.4/com/codacy/client/bitbucket/WSWrapper.scala
@@ -1,20 +1,0 @@
-package com.codacy.client.bitbucket
-
-import com.ning.http.client.AsyncHttpClientConfig
-import play.api.libs.ws.ning.{NingAsyncHttpClientConfigBuilder, NingWSClient}
-
-object WSWrapper {
-  type WSClient = play.api.libs.ws.WSClient
-  type WSRequest = play.api.libs.ws.WSRequest
-
-  def withHttpHeaders(request: WSRequest, headers: (String, String)*): WSRequest = {
-    request.withHeaders(headers: _*)
-  }
-
-  def build(): WSClient = {
-    val config = new NingAsyncHttpClientConfigBuilder().build()
-    val clientConfig = new AsyncHttpClientConfig.Builder(config).build()
-    new NingWSClient(clientConfig)
-  }
-
-}
