--- src/main/play_json_2.4/com/codacy/client/bitbucket/client/BitbucketClient.scala
@@ -1,8 +1,0 @@
-package com.codacy.client.bitbucket.client
-
-import com.codacy.client.bitbucket.WSWrapper
-import com.codacy.client.bitbucket.WSWrapper.WSClient
-import com.codacy.client.bitbucket.client.Authentication.Credentials
-
-class BitbucketClient(client: WSClient = WSWrapper.build(), credentials: Credentials)
-    extends BitbucketClientBase(client, credentials)
