+++ src/main/scala/com/codacy/client/bitbucket/client/BitbucketClient.scala
@@ -1,8 +1,7 @@
 package com.codacy.client.bitbucket.client
 
 import akka.stream.Materializer
-import com.codacy.client.bitbucket.WSWrapper
-import com.codacy.client.bitbucket.WSWrapper.WSClient
+import WSWrapper.WSClient
 import com.codacy.client.bitbucket.client.Authentication.Credentials
 
 class BitbucketClient(materializer: Materializer)(
