+++ src/main/scala/com/codacy/client/bitbucket/client/BodyWriteableWrapper.scala
@@ -1,4 +1,4 @@
-package com.codacy.client.bitbucket
+package com.codacy.client.bitbucket.client
 
 object BodyWriteableWrapper {
   type BodyWriteable[A] = play.api.libs.ws.BodyWritable[A]
