+++ src/main/scala/com/codacy/client/bitbucket/client/DefaultBodyWritables.scala
@@ -1,3 +1,3 @@
-package com.codacy.client.bitbucket
+package com.codacy.client.bitbucket.client
 
 object DefaultBodyWritables extends play.api.libs.ws.DefaultBodyWritables with play.api.libs.ws.JsonBodyWritables
