+++ src/main/scala/com/codacy/client/bitbucket/client/PageRequest.scala
@@ -8,22 +8,28 @@ class PageRequest private (cursorValue: Option[String]) {
 
 object PageRequest {
 
-  def apply(cursor: Option[String], apiBaseUrl: String): Either[String, PageRequest] = {
+  def apply(
+      cursor: Option[String],
+      apiBaseUrl: String = BitbucketClientBase.apiBaseUrl
+  ): Either[String, PageRequest] = {
+
     cursor match {
       case None => Right(new PageRequest(None))
       case Some(cursorValue) =>
         for {
-          cursorURI <- Try(java.net.URI.create(cursorValue)).toOption
-            .toRight(s"""Value "$cursorValue" is not a valid cursor: Malformed URL""")
-          apiURI <- Try(java.net.URI.create(apiBaseUrl)).toOption
-            .toRight(s"""Invalid api url $apiBaseUrl""")
+          cursorURI <- validateURI(cursorValue, s"""Value "$cursorValue" is not a valid cursor: Malformed URL""")
+          apiURI <- validateURI(apiBaseUrl, s"""Invalid api url $apiBaseUrl""")
           validationResult <- Either
             .cond(
               test = cursorURI.getHost == apiURI.getHost,
               right = new PageRequest(cursor),
-              left = s"""Value "$cursorValue" is not a valid cursor: Hostname does not match Bitbucket Server API"""
+              left = s"""Value "$cursorValue" is not a valid cursor: Hostname does not match Bitbucket Cloud API"""
             )
         } yield validationResult
     }
+  }
+
+  private def validateURI(uri: String, errorMessage: String): Either[String, java.net.URI] = {
+    Try(java.net.URI.create(uri)).toOption.toRight[String](errorMessage)
   }
 }
