+++ src/main/scala/com/codacy/client/bitbucket/v2/service/IssueServices.scala
@@ -2,7 +2,7 @@ package com.codacy.client.bitbucket.v2.service
 
 import java.net.URLEncoder
 
-import com.codacy.client.bitbucket.DefaultBodyWritables._
+import com.codacy.client.bitbucket.client.DefaultBodyWritables._
 import com.codacy.client.bitbucket.client.{BitbucketClient, Request, RequestResponse}
 import com.codacy.client.bitbucket.v2.Issue
 import play.api.libs.json.{JsString, Json}
